+++ src/test/java/org/opensearch/knn/index/codec/derivedsource/DerivedSourceSegmentAttributeParserTests.java
@@ -17,9 +17,9 @@ import java.util.Map;
 
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
-import static org.opensearch.knn.index.codec.derivedsource.DerivedSourceSegmentAttributeHelper.DERIVED_SOURCE_FIELD;
+import static org.opensearch.knn.index.codec.derivedsource.DerivedSourceSegmentAttributeParser.DERIVED_SOURCE_FIELD;
 
-public class DerivedSourceSegmentAttributeHelperTests extends KNNTestCase {
+public class DerivedSourceSegmentAttributeParserTests extends KNNTestCase {
 
     @Mock
     SegmentInfo mockSegmentInfo;
@@ -28,26 +28,33 @@ public class DerivedSourceSegmentAttributeHelperTests extends KNNTestCase {
         Map<String, String> fakeAttributes = new HashMap<>();
         when(mockSegmentInfo.putAttribute(any(), any())).thenAnswer(t -> fakeAttributes.put(t.getArgument(0), t.getArgument(1)));
 
-        DerivedSourceSegmentAttributeHelper.addDerivedVectorFieldsSegmentInfoAttribute(
+        DerivedSourceSegmentAttributeParser.addDerivedVectorFieldsSegmentInfoAttribute(
             mockSegmentInfo,
             List.of("test", "test2.nested", "vector")
         );
         assertEquals("test,test2.nested,vector", fakeAttributes.get(DERIVED_SOURCE_FIELD));
         fakeAttributes.remove(DERIVED_SOURCE_FIELD);
 
-        DerivedSourceSegmentAttributeHelper.addDerivedVectorFieldsSegmentInfoAttribute(mockSegmentInfo, List.of("test"));
+        DerivedSourceSegmentAttributeParser.addDerivedVectorFieldsSegmentInfoAttribute(mockSegmentInfo, List.of("test"));
         assertEquals("test", fakeAttributes.get(DERIVED_SOURCE_FIELD));
+
+        DerivedSourceSegmentAttributeParser.addDerivedVectorFieldsSegmentInfoAttribute(mockSegmentInfo, List.of("", "", "", "", ""));
+        assertEquals(",,,,", fakeAttributes.get(DERIVED_SOURCE_FIELD));
     }
 
     public void testParseDerivedVectorFields() {
         Map<String, String> fakeAttributes = new HashMap<>();
         when(mockSegmentInfo.getAttribute(any())).thenAnswer(t -> fakeAttributes.get(t.getArgument(0)));
 
-        assertEquals(Collections.emptyList(), DerivedSourceSegmentAttributeHelper.parseDerivedVectorFields(null));
-        assertEquals(Collections.emptyList(), DerivedSourceSegmentAttributeHelper.parseDerivedVectorFields(mockSegmentInfo));
+        assertEquals(Collections.emptyList(), DerivedSourceSegmentAttributeParser.parseDerivedVectorFields(null));
+        assertEquals(Collections.emptyList(), DerivedSourceSegmentAttributeParser.parseDerivedVectorFields(mockSegmentInfo));
 
         fakeAttributes.put(DERIVED_SOURCE_FIELD, "test,test2.nested,vector");
         List<String> expectedFieldInfos = Arrays.asList("test", "test2.nested", "vector");
-        assertEquals(expectedFieldInfos, DerivedSourceSegmentAttributeHelper.parseDerivedVectorFields(mockSegmentInfo));
+        assertEquals(expectedFieldInfos, DerivedSourceSegmentAttributeParser.parseDerivedVectorFields(mockSegmentInfo));
+
+        fakeAttributes.put(DERIVED_SOURCE_FIELD, ",,,,");
+        expectedFieldInfos = Arrays.asList("", "", "", "", "");
+        assertEquals(expectedFieldInfos, DerivedSourceSegmentAttributeParser.parseDerivedVectorFields(mockSegmentInfo));
     }
 }